[439359] Modifications to the local search project for local search debugger
This change introduces an ILocalSearchAdapter callback interface.
Through this interface the debugger (or basically any other desired
adapter) can receive information about the execution of the current
search plan.
An implementation of the ILocalSearchAdapter is also provided under the
name LocalSearchDebugger.
For testing and itroductory purposes eclipse commands are created to
start, step and run an example matching process. 

Change-Id: I44ce6d79243d9eec0116662414d9494e40c9c7a3
Signed-off-by: Marton Bur <marton.bur@gmail.com>