[419114] Clarifying handling of ambiguous type definitions
In case of conflicting type constraints (e.g. the required types have no
common subtype), an error is thrown. To make the type inferrer know
about this common subtype, a new package import can be added. Adding
such packages might require additional dependencies on the classpath; to
avoid that, from now on, this kind of issue is only a warning. 

Change-Id: Iec8c7e519c47842e5c3685e61bd4e868a9b26df2
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>