[398769] Enable all Xbase-supported number literals
By using the literal parser from the Xbase-specific NumberLiterals class
various literal types (e.g. long or float) can also be supported in the
pattern language, including the advanced features, such as the use of
'_' characters to separate parts of the numbers.

The IntValue and DoubleValue types had to be removed from the language
definition to ensure the instances can be parsed; however, they were
only used in parsing-related aspects; no runtime code had to be changed
to support this case.

In LiteralAndComputationTypeTest a few test cases were added that
illustrate the capabilities of the new features; other tests were
updated not to rely on the IntValue definitions from the pattern
language EMF model.

Change-Id: I6225b2e752d64ecadad6c8f76b226d58d5da0310
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>