[506071] Removed unnecessary dependency from maven configuration
It is not required to add the recipe project as dependency to the Maven
compiler in the Rete visualizer project, as the runtime already contains
this project, and in Maven dependent projects are already added to the
classpath. Furthermore, as p2 dependencies were not ignored, this caused
very strange issues with Xtext 2.9 where it was possible that Guava
versions 14 and 15 are mixed with each other.