[507459] Backend factory API simplified
Implementation of the hybrid pattern matching requires the provision of
API for the backend which is currently not available on the
IQueryBackendFactory.create method. To ease future changes of this API,
the IQueryBackendContext interface is introduced which aggregates all
other context provided by the engine to the backend instances.

This change involves an API break for the IQueryBackendFactory
implementations, which is permitted considering it is an internal API
between plug-ins with strict version dependencies.

Change-Id: I71abbbf311ae64b5b71dd51a5a5851332b2661e3
Signed-off-by: Grill Balazs <balazs.grill@incquerylabs.com>