[releng] Made maven artifacts compatible with Guava 15
The Xtext 2.9.2 specifies it does not work with Guava 15.0, the one used
in most Eclipse instances nowadays; however it works correctly with this
version in IDE environment. This change tries to overcomes this issue by
using Xtext 2.10 for code generation. Regardless, this does not make
Xtext 2.10 required in the runtime.

Change-Id: Ib6b211c5c081ab791d776c75e748b2f9072ad87b
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>