[507459] User-provided hints delegation fix
If the ViatraQueryEngine instantiates the result providers with already
overridden hints, it disables hybrid matching as the result provider
considers the merged hints (with engine defaults, query hints and user
hints) as the user hints, causing settings from the engine defaults to
override query hints for called patterns. The merged hints shall only be
produced by the result provider internally.

Change-Id: I7b238f9e0c786eac8e437f0b220d7617f4e78ec8
Signed-off-by: Grill Balazs <balazs.grill@incquerylabs.com>