[493616] Different handling for files not in source folders
The duplicate validation and the code generation settings should handle
differently the patterns from non-source folders (e.g. in target or bind
folder): they are not mentioned in the duplicate validator, and the
builder also does not execute on them.

As a side-effect, this also fixed an issue wrt renaming a vql file: when
the file was renamed, the old uri was still visible for duplicate
validation, and was not re-evaluated later (only in the editor). Given a
deleted file is not available in a source folder, it gets ignored here
as well.

Furthermore, bug 505616 seems also been fixed by this change as the
false paths are also excluded by this change.

Change-Id: I20dcd01a8aac250a3f20e7059239ae5a72e28b82
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>