[512759] Disable out-of-scope constraint elimination in Rete with a hint
Type constraint elimination could change the semantics of a graph
pattern in case of incomplete scope (e.g. a scope reduced to a set of a
model where there are some proxies). To disable this process, a new,
Rete-specific hint was introduced that disallows reduction for edges
that might lead out of the scope

Change-Id: Ie84e04cab7440351b306d623167a9b31128eeabc
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>
Signed-off-by: Grill Balazs <balazs.grill@incquerylabs.com>