[472850] Updated internal state handling of local search debugger
Previously, internal state of the local search debugger was managed by
commands, making it problematic to handle real uses, such as opening
multiple sessions one after the other, or reopening previously closed
ones.

This change updates the code not to provide stateless commands, while
the debugger view manages the internal state of the debugger as well.
This is still far from ideal, but makes the internal implementation much
more robust.

Change-Id: I6417b65844a22de2799305da8e4ea555e5ec7e5d
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>