[518356] Avoid duplicate statistics entries in Base index
When using statistics mode as a wildcard level followed by indexing of
certain types, it seems that count values doubled.

After further analysis, the test cases only show that EDataType indexes
are affected (why?), but for the sake of consistency, a cleanup
functionality for the statistics store was added for all types.

Change-Id: I4e266721ae84bce718f6744f89059dc77e213c7b
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>