[521758] Tests for vql generator options and fixed incorrect elements
A few Xtext-supported generator tests were added to test the new code
generation options; and a few validator updates were necessary to make
sure everything works as expected.

To execute the tests correctly, a few environment updates were also
necessary:

  * When running UI and non-UI tests in the same environment, some UI
tests failed consistently. A more detailed analysis showed that this bug
only required a following sequence: (1) non-UI test; (2) a UI test (one
of the content assist test files); (3) another non-UI test; (4) the new
generator tests. This issue could be worked around by moving all UI
tests to their separate project; in long term this is favorable as well.
  * The generator tests require a correct target platform setup that
Tycho does not correctly, see
https://bugs.eclipse.org/bugs/show_bug.cgi?id=343156 During working
around the issue, a mars-based target platform was also defined and the
new UI test project already uses it. Technically this might not be
necessary to run the tests, given we will move to Mars as a minimum
requirement in a few months, this requirement was not reverted.

Change-Id: Ia3b85314cfcb7d1ca3d4d9628dd2f7ed0995552f
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>