[467439] VIATRA Query API uses unchecked exceptions
This change creates a new exception class ViatraQueryRuntimeException,
and retrofits the existing exception classes in the Query API to rely on
this new features. This class is a runtime exception, making all
exceptions originating from VIATRA unchecked.

This allowed some cleanup in the codebase: simple repackaging of
exceptions are almost always unnecessary; and some internal uses could
be simplified this way.

Additionally, throws declarations were reviewed and updated the
following way:
 * All throws declarations of runtime exceptions were removed as they
are not enforced in any way.
 * @throws tags in Javadoc were updated in public APIs to mention the
new ViatraQueryRuntimeException class to check
 * The exceptions are not mentioned for internal APIs.

Change-Id: Iaa0f1b6a0639251fcc4c7f5c61e91cec86604bd3
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>