[527371] Soften Xtext dependency to allow Xtext 2.12 as a base
Given that Xtext 2.13 is not available on Oxygen.2, installing VIATRA
2.0 on an Oxygen Modeling Distribution is quite problematic, as there is
a version of EMF Parsley installed by default that is incompatible with
Xtext 2.13.

To allow VIATRA 2.0 to be easily installable in the latest Modeling
Distribution, we should soften our requirement on Xtext 2.12. This
should be relatively painfree for the project for now.

Change-Id: If8dcbf831f9b90eed7cf2ec7d683fe5addab0ce1
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>