[521816] Added IMultiLookup facility
The introduced multimap-like classes will be adaptable to a number of
circumstances in Rete, such as projection indexers, delta/mailbox
indexers; as well as several potential uses in the base index.

MultiLookups can map keys to a bucket of values (currently sets and
multisets are supported, delta sets to come), and support
primitive-typed (currently long only) keys and values in the low-level
storage (but with uniform API).

Change-Id: I0074ffb4f7fd640514700cce51fa16faf23b39bb
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=521816
Signed-off-by: bergmann <bergmann@mit.bme.hu>