[527641] Allow PProblems to store error locations
By storing explicit position information about error locations in
PProblems it is possible for clients of the standalone API to report
better error messages.

This feature is not used by the generated API as the generated code is
not created in case of errors in the original patterns, thus there is no
specific information available for code generation.

Change-Id: I7f98f395508be6d9a1e7854e77c72a679dbae3f8
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>