[530701] Pure whitelist registry now relies on Java ServiceLoader
Pure whitelist loading is now handled by Java Service loaders;
compatibility with previous extension-based loading is handled by a
specific service instance. In addition to this feature, the registry is
now loaded on first use; it is not necessary to call the extensionloader
manually (but still possible with a dedicated method on the
PureWhitelist class.

Given Java Service loaders is classpath-dependent, in addition to this
new extension method the extension-based is still required for plug-in
environments; but it was simplified to provide the same implementation
classes as the new solution. The language tests are executed in an
Eclipse environment, thus the corresponding test injector needed to be
updated as well.

Change-Id: Idce51abb00e58e6e7791c645c9c62e8c19a0841c
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>