[530727] Changed IResultProvider APIs to return streams and optionals
This change updates the result provider API to return a Stream of
results instead of Collections. As a related change, the aggregators was
updated with an additional method to support Mutable Aggregation as
supported by the Java 8 Stream APIs.

The local search implementation is not entirely lazy, as it can result
in multiple matchers being executed for the same pattern simultaneously
that causes errors. This issue can be revisited after fixing
https://bugs.eclipse.org/bugs/show_bug.cgi?id=528377.

Change-Id: Ia02b136eac9f72cb3de5e1239a7121426e471470
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>