[531132] Robustness update on EVM during dispose
If the Viatra Query Engine was disposed before the query event source
for EVM, the disposal of the EVM instance returns an exception because
the match update listeners cannot be removed.

Given that disposal means we are not interested in the results of the
EVM anymore, this change should not cause any issues for other users.

Change-Id: I5542807f69d1a3959f38bd2d815e1e6cc05a1d06
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>