[531193] First version of a revamped PatternParser utility
The PatternParsingUtil was deprecated in favor of a new Builder class in
PatternParser; and a PatternParser instance is expected to manage its
own state (e.g. the ResourceSet and SpecificationBuilder is reused
between different executions of the parse method).

What is missing a more fine-grained reuse option for libraries, e.g. it
should be possible to reuse parsed pattern libraries but forgetting the
previously built patterns and/or query specifications.

Change-Id: If18a0129885d5143a19f62f4e23bb1ad73f4854d
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>