[530701] Provide annotations with Java service loaders as well
In order for the annotation validation to work in maven compiler, the
corresponding language plug-ins were moved into the 'core' parent, as
they are compiled in a separate round and are available from the maven
repository as well. The viewers plug-in should also be moved there, but
before that we should consider graduating that as it would make the
build process much simpler.

Change-Id: I60df805bf7ca1dc31dcaaac7282dd93d3391ba87
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>