[532238] Refactored ILocalSearchAdapter
The interface was extended to always provide explicit information about
the current operation and do not send details about the executor.

This change includes a new interface for LS adapters that provide more
information about the plan being executed, but reduces access to the
internal execution structure.

The new interface broke all existing adapters (profiler and UI
debugger); I tried to update all of them to the new API with the
following results:

 1. The profiler was updated - the new version is much simpler and
should work as expected.
 2. The UI debugger is still basically broken. It is capable of
displaying the search plans, but is incapable of displaying any concrete
information about the execution status, e.g. what has been executed,
what is the current operation, run to breakpoints.
 3. During testing, a new adapter called ExecutionLoggerAdapter was
created that logs out the detailed execution information about the local
search.

Given the state of the UI debugger, this change removes it from the SDK
extensions feature temporarily, until everything works again. On the
other hand, the updated adapter API should be included in release 2.0,
as it allows other uses.

Change-Id: I54b3b84bf7b4a4827056f9acfe6593a5ba217a39
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>