[534435] Ensures duplicate checking of Xbase relies on source folders
When creating a query project with Gradle, the query files can only be
copied into the jar archives by adding them as resources. However, this
means that the files are copied into the build/resources folder of the
project, where the unique classname checker provided by Xbase finds it
and reports it as duplicates.

This change provides a custom implementation of this unique classname
checker that only reports issues when the duplicates are provided in a
source folder.

Change-Id: Iaec9de3f8e29b8d536c350d5193b2f4758c5632c
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>