[530519] Introduces default caching and search backends to hints
This commit consists of two major changes:

 (1) In addition to hints expressing explicit backend requirements,
implicit ones (default caching backend; default search backend can be
added). The `search` and `incremental` modifiers of a pattern are
translated to such hints.
 (2) In order for the ViatraQueryEngine to store default backends, the
dependencies between `query.runtime`, `query.runtime.rete` and
`query.runtime.localsearch` have been reorganized by having both the
rete and LS plugins to depend on the runtime, and the runtime use a
combined extension- (for plug-in environments) and serviceloader-based
(for headless environments) method for collecting the default backends
for the system.

Change-Id: Ib020befdd42fd1d7c89ebebb7de2cf23c4ffc08c
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>