[532061] Calculate number of successful operation executions
In its previous version, the LS profiler did not know whether the
execution of an operation was or was not successful, resulting in highly
surprising profile results where an operation had very few successful
executions.

This change extends the profiler to collect the number of successful and
failed executions separately, and providing both (and their sum) to its
output.

Change-Id: Icac42cb661b3ad132f443fa2dec41a45767dc40b
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>