[534435] Turns off cross-file duplicate checking outside source folders
When the vql files are copied into other folders, the cross-file
duplicate validation detected them and reported issues about them.
Because those files always had duplicates, the goal was to turn them
off; however the way that was done was incorrect. This change fixes that
issue, thus really turning off the validator in these cases.

Change-Id: I18a770ddef9334c72372361b9151a74e0bec9b8e
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>