Reduces number of files translated with asciidoctor
Starting with this change, files that should be built by asciidoctor
automatically have a file extension of `asciidoctor` while files that
are only necessary as included have a file extension of `adoc`. This
allows reducing the number of pointless error messages in the build, and
hopefully makes the code somewhat faster.

Change-Id: I40e5a3854d780bbf85ca7e0a5aad152f19836b18
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>