[459482] Shared ResourceSet experiment in TargetPlatformMetamodelsIndex
This change introduces a shared ResourceSet implementation inside of
TargetPlatformMetamodelsIndex with the goal of not having to reload them
between different editors.

However, as of now we don't have any infrastructure to unload such
models (e.g. on target platform changes), so the experiment is only
turned on when target platform caching is also enabled.

Change-Id: Icd16431325b6c3083ad5cf3e38be1391f5cc855b
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>