[459482] Avoids loading resources in ResourceSetMetamodelProviderService
The current version of ResourceSetMetamodelProviderService tries to
access EPackages by URI; however, if it was not loaded, it might result
in a HTTP connection based on the nsURI of the EPackage to be loaded.
This change turns off this resource loading, but still allows accessing
metamodels from the current ResourceSet if they are loaded.

Considering this provides was created for the now discontinued Xcore
Query integration, I don't believe this will have many side-effects.

Change-Id: Ie0473a99583732c5f1eaaf602615b3079148f67b
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>