[535815] Corrects label refresh in Viewers
The label refresh for edges was not working when the pattern describing
the edge had "additional" parameters: parameters that do represent
neither the source nor the edge of the pattern. A debugging session has
shown that this was caused by incorrect query generation: when
generating the list of required trace edges, all parameters of the
original pattern were considered, not only the ones required by the
annotations.

This change (1) fixes this by reading the provided trace maps to
generate constraints, and (2) extends the API to provide separate trace
maps for the Reference queries (used to create items and edges) and
Traced Queries (used to update and delete items and edges).

Change-Id: Ic1103421e7d2e08c15ff2d23d9bca29b9b90eece
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>