[517254] Reimplements insert type quick fix as semantic modification
By adding an EMF model object instead of the type name, we could rely on
Xtext/Xbase mechanisms to serialize the type names. As a result, type
imports not only considers language keywords but implicit and explicit
imports as well. More precisely, java.lang.String is inserted as String;
java.util.BigDecimal is inserted either as ^java.util.BigDecimal (if it
was not imported explicitly), or as BigDecimal (if it was already
imported).

Change-Id: I129300f0eb36c3da82b1498977bee83e2acd4f1b
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>