[536754] Ensures classpath validator detects required Java 8 runtime
In VIATRA 2.0 ClasspathValidator was not added to the validators,
turning off various classpath validation settings. This change turns it
back on, removes the now unnecessary runtime.localsearch requirement
checking and updates JDK library verification to check for Java 8 class.

Furthermore, a dedicated test was added to verify to functionality of
the new validator.

Change-Id: I284c0d6e9b944b746a934e482ba281fd4526ae9b
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>