[537374] Reintroduce support for negative number literals
When going for reusing literals from Xbase we have lost the ability to
describe negative numbers in the pattern bodies, unless inside a
check/eval expression.

This change extends the metamodel to support such elements; this
required some gymnastics in order to avoid conflicts in the grammar and
supporting both the interpretative and generated pattern constructions.

Change-Id: Ic421f035e75a4c3455480c4b4efadb854f3efbcb
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>