[538196] Updates local search profiler
In version 2.0, the LS profiler adapter was updated to use the Map.merge
method, but it was used incorrectly, resulting not counting the number
of matches but always storing number two in the match set.

Furthermore, the output plan ordering was thrown off by the PDisjunction
instances used in the query runtime - thus updating the body set to
maintain insertion order should fix the issue.

Change-Id: I3ca165845efbff1bf6ab6e2d3621f87cbf8ca74d
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>