[498747] Updates the VGQL metamodel and VQL code generator
This change simplifies the graphical editing of patterns in a few corner
cases by avoiding explicitly specifying unused variables for pattern
calls, replace index-based pattern call parameterization with a
name-based one, and makes it easier to reuse complex expressions such as
function evaluations and aggregated values.

A new CalledParameter EClass was introduced in the metamodel, that is a
specialized Reference class connected to a Parameter of a called
pattern. In such references, if no corresponding local expression is
selected, the generator inserts an unnamed variables ('_') to the calls.

The generator creates for each FunctionEvaluationValue and
AggregatedValue a dedicated variable, making it unnecessary to
explicitly create them in the graphical syntax when the same call needs
to be used multiple times in the same body.

On a related note, the "Assign call parameter" Edge creation removed
from the palette, as it could be replaced by the more general "Add
Reference" tool.


Change-Id: I8869a28ddbae045977a4aad81d572eec5e20cd6d
Signed-off-by: Mocsai Krisztian <krisztian.mocsai@incquerylabs.com>