[538742] Fixes LS planner logic bugs
(1) 
A PlanState represents a sequence of operations (operationsList) and 
caches the computed cost for this operation sequence. 
The list and the cost are initialized in the constructor. 
However, #categorizeChecks() also updates the operations list 
(by suffixing checks), but it failed to update the computed cost. 
This commit is intended to fix this bug.

In order to be able to continue the cost computation after the 
termination of the constructor, both the result (the cost) and an 
auxiliary variable (the cummulative product of branching factors) 
must be preserved to be available at the time of #categorizeChecks().

(2)
LocalSearchRuntimeBasedStrategy#currentState.getPresentExtends() selects 
EXTEND operations only. (As CHECK operations that can be performed 
after this new EXTEND constraint application are applied later, 
by #updateOperations).

The essence of the bugfix in this file is that the latter method is 
called to apply CHECKs immediately after the EXTEND is selected. 
Previously such CHECKS were only added to the plan after consulting 
the plan table (#determineIndices(), #checkInsertCondition()), 
so the cost of check operations 
(which can be non-negligible for e.g. unflattened pattern calls) 
were not taken into account when filling the table.


Change-Id: I313b3623b2e11d42a758f4d5790f9faba1b3b70f
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=538742
Signed-off-by: bergmann <bergmann@mit.bme.hu>