[539915] Updates transformation rule builder typing
The rule builders used for batch and event-driven transformations
required explicit match and matcher type declarations to work in Java,
as the builder was defined before the query specification providing the
type information was added. This posed no problem in Xtend, as the
extended type inferring capabilities could forward the type information.

This change provides an alternative rule builder that instead of loading
the specification in the `precondition` method of the builder receives
it during construction, thus providing the required type information in
a way appropriate for Java sources as well.

The test case includes a previously appropriate workaround as well
(explicitly providing the type parameters of the built rule), but no fix
is available yet, as it requires a big rewrite of the typing rules to
avoid the necessity of these type parameters.

Change-Id: Icdda8ea95de9e3230ec3bd2668fd4febe5ecc7bf
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>