[538847] Moves IModelConnector to a new plug-in
The IModelConnector interface provides a reusable block for accessing
models from any given Eclipse editor that was originally part of the
VIATRA Query Tooling. However, some users, most notably the validation
framework might want to rely on this functionality without all the
dependencies of the tooling, most notably Xtext.

To ensure this, a new viatra.query.runtime.ui plug-in was created, and
the IModelConnector interface was moved into that project, and all uses
and implementors of the interface was updated accordingly.

Change-Id: I41fa844e3d2b56b6f9b75bd436a5332f3be214bc
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>