[540308] Minor API cleanup
Moves the new IndexHostManipulations API to its own, experimentally
shared package (see empty x-friends declaration), and added
corresponding x-friends entries for the experimental tabular APIs as
well. This resulted in a big reduction in API warnings for us, but still
makes it clear for end-users that this API is still experimental.

Change-Id: I41219e408f151a0fba37eb2b6673884aadb6ee6d
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>