[539074] Registers validation rules for EMF Validation
Although this does not start validation automatically, it can be
triggered manually, and the results are displayed on the diagram. Some
testing is still required, but this solution is more in synch with the
workflows Sirius expects, thus less likely to break.

The integration API is almost totally independent of the queries used
here, so it might be generalized as a VIATRA Validation Framework
feature if necessary, however, some testing is required before that can
be done.

Change-Id: I253893d57baf9fef39a52110413fd29c40025268
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>