[542490] Fixes fixed priority conflict resolver logic
When replacing Multimap<Integer, Activation> from Guava with
Map<Integer, Set<Activation>> for VIATRA 2.0 as priority buckets, the
map was initialized with empty collections, while the getFirstBucket
expected that no empty collection was stored in this map.

Given that collections are removed correctly when the last activation of
a bucket is fired, this only causes problems when there are no
activations for a given priority, this issue remained invisible. This
change also includes a test that reproduces this issue to avoid such
regressions in the future.

Change-Id: I7958f9b2e3390d9a872ee4733ede96ae1c3e539c
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>