[542466] Updates rule filtering for batch transformations
Batch transformations had an incorrect handling of rule filters: while
they could be set up using the factory, the filters themselves were
ignored by the transformation. This change updates the behavior to make
sure such rules work as expected:

  * EventFilter was updated to be a Java 8 Predicate
  * The EventFilter setters were updated to expect the correct generics
  * Batch transformation rule statements were fixed to use the
rule-level setting as a default, and override them if a specific filter
was provided; if necessary, EventFilter#and can be used to provide an
extension to the default rules.
  * Added tests that illustrate this behavior

Change-Id: I78ef7ff4963facfd5640c1c44e46f6fd5f213925
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>