[542979] Extends batch transformation API with access to current rules
The fact that the set of transformation rules can be accessed from the
transformation itself allows the simple definition of incremental,
data-driven transformations, e.g. it is possible to decide whether the
transformation can be fired, or fire all rules in the transformation in
a single step.

This change introduces this functionality by:
 * Adding a new method to the batch transformation that initializes a
rule group from all _current_ rules in the transformation.
 * Adding a set of hasMatch-like statements to the statement API.
 * Adding a set of new statements that utilize the fact that the
transformation can return a default set of rules.
 * Finally, a test was added to illustrate the usage of this
functionality.


Conflicts:
	transformation/plugins/org.eclipse.viatra.transformation.runtime.emf/src/org/eclipse/viatra/transformation/runtime/emf/transformation/batch/BatchTransformationStatements.java

Change-Id: Id09b12778fdd58b96ab8ef75c70d4d258d6b47f0
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>