[545000] Adds a set of code minings
This change adds two code minings:
 * Display the name of parameters of the called patterns (if there are
at least two of them)
 * Display inferred parameter types when not present

Given we need to find out the performance and usability aspects of this
feature, the feature is off by default and can be turned on via a
preferencee setting. We might need to fine-tune these settings, e.g.
reduce the appearance of called parameter names in trivial cases, but we
need to find out whether it is necessary or not.

Change-Id: I386a2983dc087bef037cbfa2d60f80bb604d3d23
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>