[545184] Makes error handling more robust in pattern parser
When an unexpected exception is thrown by a validator (in the current
case by a completely malformed code), an issue is added to the list of
problems without a model uri.

This change fixes the found exception, and ensures that such unhandled
exceptions are handled gracefully by the standalone parser.

Change-Id: Iedecf41d6bf58c343aa7472da4ecf5663b525785
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>