[546093] Reduce log amount for duplicate notifications
In non-strict mode, base index only should report the first occurrence
of a duplicate notification as an error, further ones should be logged
as debug messages. The implementation might report more than one error
(e.g. if there are multiple Base index instances, or both the instance
and statistics store could report it separately), but this already
solves the original issue, and it makes sense for each store to report
this issue separately.

Change-Id: Idfce8056dcb9b34b822c5347211ba5bcd043fe17
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>