[549955] Adds model root access to NavigationHelperImpl
In order to allow customizing the base index contents, two methods were
added to NavigationHelperImpl to access and change the contents of the
set of managed modelRoots: one that removes already registered model
roots and another that provides a stream of current model roots.

Both of these functions are considered internal APIs (at least for now),
later they might be added to the NavigationHelper interface as well.

Change-Id: Ic1901f14eb9e5eea549cbd34898d6698f35107f6