[552532] Fixes removals during reparsing in AdvancedPatternParser
Patterns that were removed during an update operation were not handled
and their usage resulted in getting the pattern parser infrastructure
getting into an inconsistent internal state.

This change collects the set of removed patterns before parsing and
using that to calculate a list of removed patterns correctly. However,
the used algorithms does not always differentiate between added and
updated patterns correctly - however this should not result in incorrect
behavior for the users of the API, as the specifications would be
rebuilt in both cases correctly.

Change-Id: I45407eac9c0382130a89c966ed9f7d6b442afbc2
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>