[539025] Ensures correct validator usage for CPS tests
Some cps tests use the Xtext parsing infrastructure, but their setup was
inconsistent - mostly they were relying on the default injector
registration, but the DynamicModeTests class works only if the
standalone injector was registered. However, if during the run the
previous test cases registered the normal infrastructure, in EMF
registries the validators were kept.

This change introduces a JUnit rule-based implementation of XtextRunner
that correctly sets up and disposes the Xtext infrastructure on a
per-test basis, and uses this on all tests that were relying on the
parsing infrastructure. The rule is necessary, as the runner cannot be
combined with parameterized tests.

Change-Id: Ic17ac1c859426938290ead84230c46bbdfd0ab9b
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>