[559520] Ensures type inferrer errors do not break compilation
The new validation rule added by this change ensures that type
information is considered during validation correctly to avoid a NPE
during pattern compilation.

This change is intended only as a stop-gap solution until bug
https://bugs.eclipse.org/bugs/show_bug.cgi?id=559522 is fixed and the
problematic pattern included here as a test case parses and compiles
correctly. Furthermore, this fix is intended to be backported to VIARA
2.3.1 as well.

Change-Id: I4636eaf6b1afed58ec62e3925961bbf27ebd5f0a
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>