[561344] Fixes regression in parsing private pattern calls
This change fixes a strange regression of the standalone parser not
resolving pattern calls _from_ private patterns, caused by
https://git.eclipse.org/r/c/151764/ merged before 2.3.0.M4.

The problem is caused by the fact that the package declaration of a VQL
file was not considered as a required import normalizer by
EMFPatternLanguageImportNamespaceProvider and the namespace provider
only added the namespace because "by coincidence" the inferred pattern
group Jvm Declared Type had exactly the same name as the package, thus
XImportSectionNamespaceProvider added the package name as an import
normalizer to the scope.

However, the previously mentioned change made it optional to infer the
query group type, thus removing this guaranteed provider, causing the
strange issues. This change fixes this regression by explicitly
considering the package declaration as a namespace provider, thus
avoiding the original issue entirely.

Change-Id: Id96db63baad6c5ff25db091e05c22095beb6c491
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>