[571356] Reworked composite site maintenance during build
Multiversion composites are no longer maintained as it is quite hard to
consistently maintain their contents (e.g. removing old versions from
integration and maintenance branches); on the other hand, the build
creates an appropriate /latest repository as it is simple to handle via
simple parameters.

Multiversion composites for releases will be maintained manually; for
integration and milestone builds they will be removed entirely to
simplify repository management for the future.