[574447] Avoid indexing when base index is turned off in local search
The local search result provider asked for the built search operations
to index all elements that are iterated over, however this is clearly
erroneous when the backend was initialized with the no-base option. This
change updates the indexing requirements with appropriate guard
conditions, and adds a test case that verifies the indexing being turned
off as necessary.

Change-Id: I2e35ff767a0838a5e430dd8da96245530552f2f5
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>