[571048] Fixes Input Key comparator in validator
An invalid comparator caused some hardly reproducable issues during
validation - when it happened, an exception was thrown in the background
and there were no details related to the original issue.

The test case added to the repository was used to deterministically
reproduce the issue, but it is not trivial to provide a simpler example,
both the or body and the missing parameter type declarations are
required for the reproduction.

Change-Id: I42552005176d2cd25a1147abeca0a662ef2dd833
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>