[579066] Fixes subtype suggestion for patterns with disjunctions
The type validator did not calculate correctly the most specific
superclasses for inferred types of the parameters when disjunctions were
used, thus it suggested an incorrect pattern update. This commit fixes
the issue and adds an additional test case showcasing the original
issue.

Furthermore, a type in the corresponding error message was fixed.

Change-Id: I1cc6b0db686eec15174528cb79bd198a23f7a7c0
Signed-off-by: Zoltan Ujhelyi <zoltan.ujhelyi@incquerylabs.com>