Adds type tests to runtime test framework
These tests were originally created when the suite was in the examples
repository, this change updates them to the current location, adds
required file headers and also generated a test snapshot with the
results.

It is important to note that one test case is not 100% in synch with the
current implementation, that is cleanly marked in the code (and their
results ignored).

Change-Id: I906f09416853660484f0361aae61deb18b263eb3
Signed-off-by: Zoltán Ujhelyi <zoltan.ujhelyi@incquerylabs.com>
Signed-off-by: Gábor Bergmann <gabor.bergmann@incquerylabs.com>